Fix MSVC warnings
- Fix C4244: use uint8_t for keyboard row loop variable
- Suppress C4324: alignment padding is intentional for cache-line alignment
- Remove deprecated save-always from actions/cache

+++ src/service/src/KeyboardService.cpp
@@ -81,7 +81,7 @@ grpc::Status KeyboardServiceImpl::GetState(
     std::lock_guard<std::mutex> lock(mutex_);
 
     // Get keyboard matrix state from peripheral
-    for (int row = 0; row < 10; ++row) {
+    for (uint8_t row = 0; row < 10; ++row) {
         response->add_pressed_rows(keyboard_.get_row_state(row));
     }
